+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Karafka framework changelog
 
 ## 2.1.10 (Unreleased)
+- **[Feature]** Introduce Appsignal integration for errors and metrics tracking.
 - [Enhancement] Introduce `connection.client.rebalance_callback` event for instrumentation of rebalances.
+- [Fix] do not report negative lag stored in the DD listener.
 
 ## 2.1.9 (2023-08-06)
 - **[Feature]** Introduce ability to customize pause strategy on a per topic basis (Pro).